// src/app/model/Match.ts
export class Match {
  homeTeam: string;
  homeTeamCountry: string;
  homeTeamLastNineGoals: any[];
  homeTeamHomeLastNineGoals: any[];
  awayTeam: string;
  awayTeamCountry: string;
  awayTeamLastNineGoals: any[];
  awayTeamAwayLastNineGoals: any[];
  againstEachOther: any[];
  againstEachOtherHomeAway: any[];
  time: string;
  averageTotalGoals: string;
  averageTotalGoalsChanged: string;

}